+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Add clear cache button to settings
 - Restrict local permissions for BitBoxApp config files and directories
 - Bitcoin: display zero amounts without decimal places
+- Add option to navigate to "Used addresses" in sign-message workflow
+- Floating mobile bottom navigation bar
 
 ## v4.51.3
 - Bundle BitBox02 and BitBox02 Nova firmware version v9.26.4